+++ app/Http/Controllers/SupervisorsController.php
@@ -8,6 +8,7 @@ use Auth;
 use JavaScript;
 
 use App\Supervisor;
+use App\User;
 
 use App\Http\Requests\CreateSupervisorRequest;
 use App\Http\Requests;
@@ -23,7 +24,7 @@ class SupervisorsController extends Controller
     {
         $this->middleware('auth');
     }
-    
+
     /**
      * Display a listing of the resource.
      *
@@ -31,10 +32,23 @@ class SupervisorsController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if($user->cannot('index', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator()){
+            $supervisors = $user->userable()->supervisors()->get();
+        }else{
+            $supervisors = $user->userable()->admin()->supervisors()->get();
+        }
+
         JavaScript::put([
             'click_url' => url('supervisors').'/',
         ]);
-        $supervisors = Auth::user()->supervisors;
+
         return view('supervisors.index', compact('supervisors'));
     }
 
@@ -45,6 +59,13 @@ class SupervisorsController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
+        if($user->cannot('create', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
         return view('supervisors.create', compact('supervisors'));
     }
 
@@ -56,22 +77,43 @@ class SupervisorsController extends Controller
      */
     public function store(CreateSupervisorRequest $request)
     {
+        $user = Auth::user();
+        if($user->cannot('create', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $admin = $user->userable();
+        }else{
+            $admin = $user->userable()->admin();
+        }
+
         $supervisor = Supervisor::create([
             'name'      => $request->name,
             'last_name' => $request->last_name,
-            'email'     => $request->email,
-            'password'  => $request->password,
             'cellphone' => $request->cellphone,
             'address'   => $request->address,
             'language'  => $request->language,
             'comments'  => $request->comments,
-            'user_id'   => Auth::user()->id,
+            'admin_id'   => $admin->id,
         ]);
+        $user = User::create([
+            'email' => $request->email,
+            'password' => bcrypt(str_random(9)),
+            'userable_id' => $supervisor->id,
+            'userable_type' => 'App\Supervisor',
+            'remember_token' => str_random(10),
+            'api_token' => str_random(60),
+        ]);
+
         $photo = true;
         if($request->photo){
             $photo = $supervisor->addImageFromForm($request->file('photo'));
         }
-        if($supervisor && $photo){
+        if($user && $supervisor && $photo){
             flash()->success('Created', 'New supervisor successfully created.');
             return redirect('supervisors');
         }
@@ -87,7 +129,19 @@ class SupervisorsController extends Controller
      */
     public function show($seq_id)
     {
-        $supervisor = Auth::user()->supervisorBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('show', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator()){
+            $supervisor = $user->userable()->supervisorBySeqId($seq_id);
+        }else{
+            $supervisor = $user->userable()->admin()->supervisorBySeqId($seq_id);
+        }
+
         return view('supervisors.show', compact('supervisor'));
     }
 
@@ -99,7 +153,19 @@ class SupervisorsController extends Controller
      */
     public function edit($seq_id)
     {
-        $supervisor = Auth::user()->supervisorBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('edit', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator()){
+            $supervisor = $user->userable()->supervisorBySeqId($seq_id);
+        }else{
+            $supervisor = $user->userable()->admin()->supervisorBySeqId($seq_id);
+        }
+
         return view('supervisors.edit', compact('supervisor'));
     }
 
@@ -112,12 +178,25 @@ class SupervisorsController extends Controller
      */
     public function update(CreateSupervisorRequest $request, $seq_id)
     {
-        $supervisor = Auth::user()->supervisorBySeqId($seq_id);
+        $logged_user = Auth::user();
+        if($logged_user->cannot('edit', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($logged_user->isAdministrator()){
+            $supervisor = $logged_user->userable()->supervisorBySeqId($seq_id);
+        }else{
+            $supervisor = $logged_user->userable()->admin()->supervisorBySeqId($seq_id);
+        }
+        $user = $supervisor->user();
+
+        $user->email = $request->email;
+        $user->password = $request->password;
 
         $supervisor->name = $request->name;
         $supervisor->last_name = $request->last_name;
-        $supervisor->email = $request->email;
-        $supervisor->password = $request->password;
         $supervisor->cellphone = $request->cellphone;
         $supervisor->address = $request->address;
         $supervisor->language = $request->language;
@@ -129,7 +208,7 @@ class SupervisorsController extends Controller
             $photo = $supervisor->addImageFromForm($request->file('photo'));
         }
 
-        if($supervisor->save() && $photo){
+        if($user->save() && $supervisor->save() && $photo){
             flash()->success('Updated', 'New supervisor successfully updated.');
             return redirect('supervisors/'.$seq_id);
         }
@@ -145,7 +224,19 @@ class SupervisorsController extends Controller
      */
     public function destroy($seq_id)
     {
-        $supervisor = Auth::user()->supervisorBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('destroy', Supervisor::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator()){
+            $supervisor = $user->userable()->supervisorBySeqId($seq_id);
+        }else{
+            $supervisor = $user->userable()->admin()->supervisorBySeqId($seq_id);
+        }
+
         if($supervisor->delete()){
             flash()->success('Deleted', 'The supervisor was successfuly deleted');
             return redirect('supervisors');
